Migrates away from Durable Objects for logging
