+++ apps/server/wrangler.jsonc
@@ -58,10 +58,7 @@
             "name": "WORKFLOW_RUNNER",
             "class_name": "WorkflowRunner",
           },
-          {
-            "name": "LOGGING",
-            "class_name": "LoggingDurableObject",
-          },
+
           {
             "name": "THREAD_SYNC_WORKER",
             "class_name": "ThreadSyncWorker",
@@ -148,10 +145,7 @@
           "tag": "v9",
           "new_sqlite_classes": ["ShardRegistry"],
         },
-        {
-          "tag": "v10",
-          "new_classes": ["LoggingDurableObject"],
-        },
+
       ],
 
       "observability": {
@@ -293,10 +287,7 @@
             "name": "SHARD_REGISTRY",
             "class_name": "ShardRegistry",
           },
-          {
-            "name": "LOGGING",
-            "class_name": "LoggingDurableObject",
-          },
+
         ],
       },
       "workflows": [
@@ -385,10 +376,7 @@
           "tag": "v10",
           "new_sqlite_classes": ["ShardRegistry"],
         },
-        {
-          "tag": "v11",
-          "new_classes": ["LoggingDurableObject"],
-        },
+
       ],
       "observability": {
         "enabled": true,
@@ -533,10 +521,7 @@
             "name": "SHARD_REGISTRY",
             "class_name": "ShardRegistry",
           },
-          {
-            "name": "LOGGING",
-            "class_name": "LoggingDurableObject",
-          },
+
         ],
       },
       "workflows": [
@@ -619,10 +604,7 @@
           "tag": "v10",
           "new_sqlite_classes": ["ShardRegistry"],
         },
-        {
-          "tag": "v11",
-          "new_classes": ["LoggingDurableObject"],
-        },
+
       ],
       "vars": {
         "NODE_ENV": "production",
